cast gbps result to float64

+++ traffic_monitor/cache/cache.go
@@ -155,7 +155,7 @@ func ComputedStats() map[string]StatComputeFunc {
 			return info.Vitals.KbpsOut
 		},
 		"gbps": func(info ResultInfo, serverInfo tc.TrafficServer, serverProfile tc.TMProfile, combinedState tc.IsAvailable) interface{} {
-			return info.Vitals.KbpsOut / 1000000
+			return float64(info.Vitals.KbpsOut) / 1000000.0
 		},
 		"loadavg": func(info ResultInfo, serverInfo tc.TrafficServer, serverProfile tc.TMProfile, combinedState tc.IsAvailable) interface{} {
 			return info.Vitals.LoadAvg
